+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   custom serializer for a class.
 - Make it possible to merge `Serializers` instances: add a `serializers`
   getter that returns the installed serializers.
+- Allow `analyzer` 0.34.0.
 
 # 6.2.0